pyproject.toml
[project]
name = "my-langchain-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]


[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # Errors
    "F",   # Pyflakes
    "I",   # Import sorting
    "W",   # Warnings
]

[tool.ruff.format]
quote-style = "double"
